+++ feature/search/src/main/java/com/sumin/feature/search/HospitalItemUiState.kt
@@ -1,0 +1,9 @@
+package com.sumin.feature.search
+
+data class HospitalItemUiState(
+    val id: String,
+    val code: String,
+    val hospitalName: String,
+    val sidoAddr: String,
+    val sgguAddr: String
+)
