+++ feature/search/src/main/res/layout/fragment_search.xml
@@ -11,22 +11,20 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:padding="18dp">
+        android:layout_height="match_parent">
 
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tool_bar"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
+            android:layout_margin="16dp"
             app:contentInsetStart="0dp"
             app:layout_constraintTop_toTopOf="parent">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:paddingStart="14dp"
-                android:paddingEnd="14dp">
+                android:layout_height="match_parent">
 
                 <androidx.appcompat.widget.SearchView
                     android:id="@+id/search_view"
@@ -50,9 +48,21 @@
             android:id="@+id/rv_hospital_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_marginTop="16dp"
+            android:layout_margin="16dp"
             app:layout_constraintTop_toBottomOf="@id/tool_bar"
             app:layout_constraintBottom_toBottomOf="parent"/>
+
+        <ProgressBar
+            style="?android:attr/progressBarStyleHorizontal"
+            android:id="@+id/progress_bar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:indeterminate="true"
+            android:indeterminateTint="#03A9F4"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/tool_bar" />
+
 
     </androidx.constraintlayout.widget.ConstraintLayout>
